+++ client/src/services/analytics.js
@@ -34,7 +34,7 @@ const pageView = (pageName, pageCat = null) => {
   });
 };
 
-const trackEvent = (eventName, properties={}) => {
+const trackEvent = (eventName, properties = {}) => {
   if (!enabled()) {
     return;
   }
@@ -42,8 +42,4 @@ const trackEvent = (eventName, properties={}) => {
   window.analytics.track(eventName, properties);
 };
 
-export {
-  init,
-  pageView,
-  trackEvent,
-};
+export { init, pageView, trackEvent };